tests: organize providers playwright tests
