+++ tests/package.json
@@ -12,7 +12,7 @@
     "test:jupyterlab": "playwright test jupyterlab_ui_tests --retries=3 --project=\"chromium\"",
     "test:notebook": "playwright test notebook_ui_tests --retries=3 --project=\"chromium\"",
     "test:mitoai": "playwright test mitoai_ui_tests --retries=3 --workers=3 --project=\"chromium\"",
-    "test:mitoai:gemini": "playwright test mitoai_ui_tests/gemini.spec.ts --retries=3 --workers=3 --project=\"chromium\"",
+    "test:llm-providers": "playwright test llm_providers_tests --retries=3 --workers=3 --project=\"chromium\"",
     "test:mito-sql-cell": "playwright test mito_sql_ui_tests --retries=3 --project=\"chromium\""
   },
   "author": "",
